renamed for uniformity and removed unnecessary files
