+++ src/layers/olympex-layers/components/index.js
@@ -1,4 +1,4 @@
-import {campaign, logo, description, dois, links, legends, defaultCamera} from "./constant";
+import {campaign, logo, description, dois, links, legends, defaultCamera} from "./constantLayers";
 import instrumentLayers from "./instrumentLayers";
 
 export {campaign, logo, description, dois, links, legends, defaultCamera, instrumentLayers}
